deploy 24-02-2021

+++ src/main/java/com/main/glory/servicesImpl/ShadeServiceImpl.java
@@ -192,6 +192,10 @@ public class ShadeServiceImpl implements ShadeServicesInterface {
 		if(shadeMast.isEmpty()){
 			throw new Exception("shade data does not exist with id:"+id);
 		}
+		//check the production is avialble or not
+		List<ProductionPlan> productionPlans = productionPlanDao.getAllProductionByShadeId(id);
+		if(!productionPlans.isEmpty())
+			throw new Exception("remove the production first");
 		shadeMastDao.deleteById(id);
 
 		return true;
@@ -448,5 +452,9 @@ public class ShadeServiceImpl implements ShadeServicesInterface {
 
 	public List<ShadeData> getShadDataByItemId(Long e) {
 		return shadeDataDao.getShadeDataByItemid(e);
+	}
+
+	public List<ShadeMast> getAllShadeByProcessId(Long id) {
+		return shadeMastDao.getAllShadeByProcessId(id);
 	}
 }
